Fixed krho calculation to be consistent with eps
Previously krho and eps were separately estimated relative to GM reference levels. This is fine as long as the reference K0, eps0, and N0 are always self-consistent, but it is unecessarily redundant and prone to user error. I found, for example, that the K0 used was consistent with the eps0 from Gregg 2003 but inconsistent with the present default value from Waterman 2014, giving errors of ~15% in the final diffusivities.

I've fixed the code by implementing a separate `diffusivity` function which compute the diffusivity from the dissipation rate, the stratification, and a mixing coefficient (set to Gam=0.2 by default, as in Gregg 2003 and others). Now all one needs to do is change eps0 and the diffusivity will always be self-consistent.