improved overturn & shearstrain documentation and output conventions; fix shearstrain indexing bugs (#80)
* Added references.bib bibtex file to docs config

In `conf.py`

* Commented out custom table css to fix doc build

I had to comment out the css overwrites because I think they were completely overwriting the entire default css sheet.

Based on what it says here (https://www.sphinx-doc.org/en/master/usage/configuration.html) it seems that `CSS @import` is the correct way of doing small overwrites of the default css sheet.

* Very minor changes and typo corrections

* Added Thorpe Scale info box and fixed displacements

- Added an info box explaining the basic idea behind Thorpe Scale methods for estimating dissipation rates

- Made Thorpe displacements more intuitive by unsorting them, so that the displacement is relative to the initial unsorted profile
- Importantly, this doesn't change any calculations!
- Added the sorting indices `sidx` to the diagnostics as well
- Made some small changes to the notebook to reflect these changes

* Fixed major indexing bug in strain variance calc

There are steps in which the shear/strain spectra arrays are sub-sampled:
1) using `m_include_sh` or `m_include_st` to zoom in on the spectral window specified by the user
2) using `iimsh` or `iimst` to cutoff high wavenumbers.

Previously, the `iim*` erroneously refered to the indices of arrays that were already indexed by `m_include_*` and yet were later erroneously used to index the full arrays (e.g. `m[iimsh]`). This is typically not a problem for shear because the first index is usually 0 (the window scale) and thus `iimsh` is equal to `m_include_sh[iimsh]`. However, for strain the first index is typically strictly greater than 0, so `iimst` is almost never equal to `m_include_st[iimst]`.

Essentially what this means is that strain variance was typically being integrated over a spectral window of much larger scales than specified by the user, and was thus giving erroneous results (I am not exactly sure the theoretical reason that strain is meant to be evaluated over higher wave numbers than shear, but Kunze recommends a min wavenumber that corresponds to a max wavelength of 150 m. In the documentation example, this was erroneously being set to 300 m by this bug). This bug was not as bad as it could have been because it was applied to both the GM and observed spectra, and so the errors cancelled out to some extent.

* Modularized dissipation calculation and added diags

- I made the formulas for calculating the final dissipation rates into their own functions (one for shearstrain and one for strain-only). I also explicitly related the diffusivity expression to the dissipation rate function, rather than hard-coding both.

- I added diags for Nm and the corrected shear/strain ratio (with non-physical values of Rw<1 set to 1.01) which both should probably be accessible to users for basic quality control and data exploration.

* Fixed krho calculation to be consistent with eps

Previously krho and eps were separately estimated relative to GM reference levels. This is fine as long as the reference K0, eps0, and N0 are always self-consistent, but it is unecessarily redundant and prone to user error. I found, for example, that the K0 used was consistent with the eps0 from Gregg 2003 but inconsistent with the present default value from Waterman 2014, giving errors of ~15% in the final diffusivities.

I've fixed the code by implementing a separate `diffusivity` function which compute the diffusivity from the dissipation rate, the stratification, and a mixing coefficient (set to Gam=0.2 by default, as in Gregg 2003 and others). Now all one needs to do is change eps0 and the diffusivity will always be self-consistent.

* Added Int_sh_gm and Int_st_gm as diagnostics

This are the final diagnostics required for users to recreate the final dissipation profile based on profiles of each individual term (stratification, shear or strain variance, and correction terms for aspect ratio and Coriolis frequency).

* Finalized updated shearstrain docs

- Includes an info panel the summarizes the theory and practice of internal wave finescale paramterization
- Includes a basic explanation of the integration window parameters and prints their values in physical wavelength units
- Includes all of the bug fixes in earlier commits
- Includes plots of more granular diagnostics, both in terms of spectra for a given depth window and also for full vertical profiles
- Uses more intuitive `plt.stairs` plotting convention than misleading `plt.steps`
- Declutters some of the plotting code

* Fixed formatting of latex equations

* add docs for new shearstrain functions

- add docstring for diffusivity
- add diffusivity, aspect_ratio_correction_shst, aspect_ratio_correction_st to api

* Clean up shearstrain tutorial notebook

Co-authored-by: Gunnar Voet <gvoet@ucsd.edu>